windows_requirements.txt
# Windows UI Automation
pywinauto==0.6.9
pywin32==311
pyautogui==0.9.54

# Image Processing
Pillow==12.1.0

# HTTP Requests & API clients
requests==2.32.5
groq==1.0.0
openai==2.17.0
anthropic==0.83.0
google-genai==1.64.0

# Environment & Configuration
python-dotenv==1.2.1
pywebview==6.1

# Kernel-level Input (for UAC interaction)
interception-python

opencv-python
numpy==2.4.2
keyboard
mss
flask==3.1.2
psutil

#ocr
winrt-runtime
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Media.Ocr
winrt-Windows.Globalization
winrt-Windows.Graphics.Imaging
winrt-Windows.Storage.Streams